backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
requests
torch
torchvision
Pillow
